Make WinToast implementation private
Make the WinToast class implementation private so we can remove all the
external dependencies on Windows 10 SDK in the public header.
